refactor: rename NovixBottomSheet to ModalBottomSheet
The `NovixBottomSheetState` has been renamed to `SheetState` for clarity.

Additionally, the calculation for the maximum height of the filter bottom sheet now uses `LocalWindowInfo.current.containerSize.height` instead of `LocalConfiguration.current.screenHeightDp`.
