+++ designSystem/src/main/java/com/london/designsystem/component/BottomSheet.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.material3.BottomSheetDefaults
 import androidx.compose.material3.ModalBottomSheetDefaults
-import androidx.compose.material3.SheetState
 import androidx.compose.material3.SheetValue
 import androidx.compose.material3.contentColorFor
 import androidx.compose.material3.rememberModalBottomSheetState
@@ -17,7 +16,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.london.designsystem.theme.NovixTheme
 
-class NovixBottomSheetState(val sheetState: SheetState) {
+class SheetState(val sheetState: androidx.compose.material3.SheetState) {
     suspend fun show() = sheetState.show()
     suspend fun hide() = sheetState.hide()
     val isVisible: Boolean get() = sheetState.isVisible
@@ -30,15 +29,15 @@ class NovixBottomSheetState(val sheetState: SheetState) {
 @Composable
 fun rememberNovixModalBottomSheetState(
     skipPartiallyExpanded: Boolean = true
-): NovixBottomSheetState {
+): SheetState {
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = skipPartiallyExpanded)
-    return remember { NovixBottomSheetState(sheetState) }
+    return remember { SheetState(sheetState) }
 }
 
 @Composable
-fun NovixBottomSheet(
+fun ModalBottomSheet(
     onDismissRequest: () -> Unit,
-    state: NovixBottomSheetState,
+    state: SheetState,
     modifier: Modifier = Modifier,
     sheetMaxWidth: Dp = BottomSheetDefaults.SheetMaxWidth,
     shape: Shape = BottomSheetDefaults.ExpandedShape,
