+++ presentation/src/main/java/com/london/presentation/screen/search/SearchScreen.kt
@@ -62,8 +62,8 @@ import com.london.presentation.composables.ActorsLayout
 import com.london.presentation.composables.MoviesLayOut
 import com.london.presentation.composables.TriangleBlurredShape
 import com.london.presentation.composables.TvShowLayOut
-import com.london.presentation.composables.filterbottomsheet.FilterBottomSheet
-import com.london.presentation.composables.filterbottomsheet.FilterState
+import com.london.presentation.composables.FilterBottomSheet
+import com.london.presentation.composables.FilterState
 import com.london.presentation.utils.Listen
 import com.london.presentation.utils.ResultOrEmpty
 import org.koin.androidx.compose.koinViewModel
